+++ CHANGELOG.md
@@ -49,9 +49,9 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add functionality to automatically delete expired orders - #12710 by @korycins
 - Handle error raised by 0Auth when fetching token - #12672 by @IKarbowiakg
 - Fix adding new lines to draft order when existing line has deleted product - #12711 by @SzymJ
-- Add `price_override` to `OrderLine` for set custom price in mutation `OrderLineCreateInput` - #12855 by @KirillPlaksin
 - Upgrade checkout `complete_checkout` to assign guest checkout to account if exists - #12758 by @FremahA
 - Remove `ENABLE_ACCOUNT_CONFIRMATION_BY_EMAIL` env variable from settings - ##12781 by @Szym
+- Allow defining a custom price in draft orders - #12855 by @KirillPlaksin
 
 # 3.13.0